+++ lib/modules/style_manager/library/library_style_manager.dart
@@ -1,11 +1,5 @@
 library library_style_manager;
 
-//bloc
-export 'package:chat_app/modules/style_manager/bloc/change_theme_bloc/change_theme_bloc.dart';
-export 'package:chat_app/modules/style_manager/bloc/change_theme_bloc/change_theme_event.dart';
-export 'package:chat_app/modules/style_manager/bloc/change_theme_bloc/change_theme_state.dart';
-
-//themes
