+++ lib/ui/pages/settings_page/settings_page.dart
@@ -1,9 +1,4 @@
-import 'package:chat_app/modules/style_manager/bloc/change_theme_bloc/change_theme_bloc.dart';
-import 'package:chat_app/modules/style_manager/bloc/change_theme_bloc/change_theme_event.dart';
-
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import '../../widgets/settings_page/theme_settings_widget.dart';
 
 class SettingsPage extends StatelessWidget {
   const SettingsPage({Key? key}) : super(key: key);
